+++ CHANGELOG.md
@@ -10,7 +10,7 @@ The format is based on [Keep a Changelog], and this project adheres to
 [keep a changelog]: https://keepachangelog.com/en/1.0.0/
 [semantic versioning]: https://semver.org/spec/v2.0.0.html
 
-## Unreleased
+## [0.7.0] - 2024-01-17
 
 ### Changed
 
@@ -172,6 +172,7 @@ placeholder format (`?` for MySQL, `$1` for everything else).
 [0.6.3]: https://github.com/dogmatiq/projectionkit/releases/tag/v0.6.3
 [0.6.4]: https://github.com/dogmatiq/projectionkit/releases/tag/v0.6.4
 [0.6.5]: https://github.com/dogmatiq/projectionkit/releases/tag/v0.6.5
+[0.7.0]: https://github.com/dogmatiq/projectionkit/releases/tag/v0.7.0
 
 <!-- version template
 ## [0.0.1] - YYYY-MM-DD